Move ModelFactory to utils directory
