+++ src/main/java/com/github/protocolfuzzing/protocolstatefuzzer/utils/ModelFactory.java
@@ -1,6 +1,5 @@
-package com.github.protocolfuzzing.protocolstatefuzzer.statefuzzer.testrunner.core;
+package com.github.protocolfuzzing.protocolstatefuzzer.utils;
 
-import com.github.protocolfuzzing.protocolstatefuzzer.utils.MealyDotParser;
 import com.github.protocolfuzzing.protocolstatefuzzer.utils.MealyDotParser.MealyInputOutputProcessor;
 import net.automatalib.automaton.transducer.CompactMealy;
 import net.automatalib.automaton.transducer.MealyMachine;
